requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
filetype==1.2.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.1.5
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.45
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
packaging==24.2
portalocker==2.10.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
qdrant-client==1.14.2
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
yarl==1.20.1
zstandard==0.23.0
